+++ ChangeLog.md
@@ -7,6 +7,7 @@
     Release date: 
 
     - Let the application configure allowed and not allowed results logs on the console
+    - Define allowedIfTaskNotFound and warnIfTaskNotFound configuration parameters
 
 ### 1.0.0